Update base for Update on "[Interp] Only return stack of call sites from Function value"
The only place where the stack is used is in the Var Non-Local rule: the stack returned from evaluating the LHS of the most recent function application is used to look up the variable from an outer scope. Thus, the stack only needs to be returned from the Function value.

Test plan:

The tests pass.

[ghstack-poisoned]