[Refactor][Tests] Optionally run benchmarks with `--bench`
This PR adds a `--bench` commandline flag to conditionally run the benchmarks
for the DDE interpreter, saving time if we only want to run unit tests.

This PR also fixes the benchmarks to truly run the interpreter on fibonnaci
programs of *different* inputs. The result of the benchmarks is consistent
with the expected linear time complexity of the interpreter when run on
exponential time algorithms.

Test plan:

`dune exec -- ./tests/tests.exe --bench` to run both the benchmarks and unit
tests. The running time of interpretation increases linearly against inputs to
functions and all tests pass.

[ghstack-poisoned]