[PA] Fix/refine Var rules
Fix the Var Non-Local rule to enumerate over all possible versions of the
parent function (of the variable) with the same label but different call
site stacks.

Thread S (the set of call site stacks) through the Var and Application
rules, instead of unioning all at once, for consistency and performance.
All rules will be updated as such in the next PR.

ghstack-source-id: 8bf93a398b5302879269e8476ca37f4a836b066a
Pull Request resolved: https://github.com/JHU-PL-Lab/dde/pull/22