Update base for Update on "[Fix][PA] Avoid false positives in stubbing function applications"
We only want to stub a function application when it really is the source
of a cycle in the program. For a program like
`((fun x -> fun y -> x + y) 1) 2`, we do not want to trigger the Stub
rule just because the inner application has already been visited when
getting to the non-local variable `x`, since this program is acyclic.

Thus, the resolution is to add an optional paramter to `analyze_aux`
that, when set to `false`, skips the Stub rule for function
applications. It defaults to `true`.

Test plan:

All tests pass, including the aforementioned one.

[ghstack-poisoned]