[Fix][PA] Avoid false positives when stubbing function applications
We only want to stub a function application when it really is the source
of a cycle in the program. For a program like
`((fun x -> fun y -> x + y) 1) 2`, we do not want to trigger the Stub
rule just because the inner application has already been visited when
getting to the non-local variable `x`, since this program is acyclic.

Thus, the resolution is to add an optional paramter to `analyze_aux`
that, when set to `false`, skips the Stub rule for function
applications. It defaults to `true`.

Test plan:

All tests pass, including the aforementioned one.

ghstack-source-id: 01537cb6147b744cbb016be2337df8a749b08d0f
Pull Request resolved: https://github.com/JHU-PL-Lab/dde/pull/26