Update on "[Fix][PA] Use immutable Set for visited program points"
The purpose of this change is to use the right visited V set for the
right expression being evaluated. A mutable set was previously used, and
that led to issues with the Stub rule being falsely triggered on
function applications evaluated in a different, earlier subtree. Since
the point of the Stub rule is to mitigate cyclic programs (on a single
path up the tree), the Stub rule being triggered is not the correct
semantics. Avoiding the re-evaluation of the same expression across
different subtrees is a matter caching, which is orthogonal to the issue
at hand.

Now, V is passed solely up its own subtree and thus this issue is
funmentally resolved.

Test plan:

All tests, including `((fun x -> fun y -> x + y) 1) 2`, pass.

[ghstack-poisoned]