Update on "[WIP][Tests][PA] Enhance and validate randomized tests"
This commit optimizes the generated ASTs as much as possible to get good
test cases. When possible, a free variable use is replaced with a use of
a random preceding definition. Fb's `eval` is then run on each generated
AST to validate that it's closed and devoid of runtime type errors.

Test case:

`dune test program_analysis`

[ghstack-poisoned]