Update on "[Tests][PA] Validate randomized tests and maximize variable use"
This commit optimizes the generated ASTs to improve test cases. When
possible, a free variable use is replaced with a use of a random
preceding definition. Fb's `eval` is then run on each generated AST to
validate that it's closed and devoid of runtime type errors.

Test case:

`dune test program_analysis`

[ghstack-poisoned]