[Formal] Auto-assign expression labels
A DDE expr can now be labeled and turned into an lexpr automatically.
It's now much more convenient to write complex programs.

To make this work, unlabeled versions of Appl and Fun are added
alongside the labeled versions in the type definition. Alternate
notations are also defined for them, like `<-*` for `LAppl` and `->*`
for `LFun`.

The drawback is that helper functions to assign labels and build maps
now have to match cases that are unreachable yet has to return values
due to Coq being purely functional. An elegant solution to this may come
in a future commit.

Test plan:

`dune build` to see that everything compiles.

[ghstack-poisoned]