Update base for Update on "[PA] `Eval` if conditions"
This commit adds an `Eval` function (`eval_int` + `eval_bool`) that
evaluates the condition of an if expression to increase the precision of
the analysis. It considers stubs as any value, which is where we lose a
bit of precision.

To make this work, we introduce a two-pass mechanism to Application Stub
so that we can correctly `Eval` the stub within the if condition before
stubbing the recursive call. This gives results that are more
informative of the actual behavior of the program. Consider the
following example:

```ocaml
let id = fun self -> fun n -> if n = 0 then 0 else 1 + self self (n - 1) in id id 10
```

computes to `(1 + (1 + (1 + (0 | (1 + stub)))))`. Without the two-pass
mechanism, we would get `(1 + (1 + (1 + stub)))` which doesn't imply the
possibility of the base case.

[ghstack-poisoned]