[Formal] Fix stack pruning algorithm
This commit fixes `prune_sigma2` so that it preserves the order of the
elements in it (if there are less than or equal to two). Specify and
prove relevant lemmas about it. Adjust `solve_analyze` to properly match
each stack in S before deciding whether to stitch with it.

Test plan:

CI passes. `dune build formal` and see that everything compiles.

[ghstack-poisoned]