[PA] Sync with theory + major refactoring
This commit adjusts the data types of the language grammar and specific
program analysis mechanics to match with the latest developments in the
written theory.

`lib.ml` was getting a bit long, so I broke it down into three files:

- `lib.ml` contains the core `analyze` function and the if-condition
Eval function.
- `grammar.ml` contains data types for the grammar of result values of
the program analysis, as well as utility modules to serve as base types
for the various set data structures used in `lib.ml`.
- `utils.ml` contains utility functions used in `lib.ml`. Note that it's
different from `debugutils.ml` in that the latter is used for
interactive debugging.

Test plan:

`dune test program_analysis` and all tests pass.

ghstack-source-id: 965feee51f9c03ad8ed836f349a391dbab36b7cd
Pull Request resolved: https://github.com/JHU-PL-Lab/dde/pull/72