Update base for Update on "[PA] Use results instead of expressions in path conditions"
This commit amends 4a656d4. Without additional info, it isn't very
helpful to show only the expression that's true/false. So, we instead
show its evaluated result. Perhaps in the future it's a better idea to
show the expression *and* annotate it with its position in the source
program.

Test plan:

`dune test program_analysis`

[ghstack-poisoned]