[PA] Use results instead of expressions in path conditions
This commit amends 4a656d4. Without additional info, it isn't very
helpful to show only the expression that's true/false. So, we instead
show its evaluated result. Perhaps in the future it's a better idea to
show the expression *and* annotate it with its position in the source
program.

Test plan:

`dune test program_analysis`

ghstack-source-id: 637a661704e6e50ebfe69e2ecf03f27ec20b6878
Pull Request resolved: https://github.com/JHU-PL-Lab/dde/pull/78