Update on "[PA] Probe CHC generation issue on recursive example"
Turns out to be due to the internal randomness of Z3/Spacer that an
absurdly verbose solution (~150k lines) is derived from the generated
CHCs (~450 lines). The solution is accurate, however, and serves our
purpose of verifying the correctness of the analysis result.

This doesn't reproduce using a local installation of Z3 or an online
playground where a concise ~100-line solution is derived with an
accurate root predicate instantiation. This is how we made sure that
the model from the automated version in OCaml is behaviorally the same,
especially on edge cases. Now chcs_of_res/atom works well for all tests,
including church numerals.

Test plan:

`dune test program_analysis`

[ghstack-poisoned]