[PA] Use Z3 to solve if conditions
Apply CHC generator to deciding which branch to take in conditionals.
Fix a bug in the generator such that a pair of labeled result atom and
stub atom are now assigned the same ID and the resulting CHCs are also
cyclic (edges are logical implications).

Test plan:

`dune test program_analysis`

ghstack-source-id: bcc61d6995a45141d0f80b4155087829f97fc849
Pull Request resolved: https://github.com/JHU-PL-Lab/dde/pull/90