Update on "Make assertions a language feature + more arithmetic operators"
Instead of specifying the CHC clause to add for each program in tests,
we introduce a language construct `letassert x = e in ...` where the
ellipsis denotes an assertion over the value of x. We can then translate
it into a CHC clause to be solved together with all else. The form of
this assertion is currently pretty restrictive but may be relaxed in
the future.

More arithmetic operators (>=, >, <=, <) are also added so it's easier
to write more expressive programs and tests.

A new flag (defaults to true) is added to the `analyze` function to
optionally verify the result using Spacer. Thus, the test framework now
perform unit testing and verification testing over analysis results
together.

Test plan:

`dune test program_analysis`

[ghstack-poisoned]