Make assertions a language feature + more arithmetic operators
Instead of specifying the CHC clause to add for each program in tests,
we introduce a language construct `letassert x = e in ...` where the
ellipsis denotes an assertion over the value of x. We can then translate
it into a CHC clause to be solved together with all else. The form of
this assertion is currently pretty restrictive but may be relaxed in
the future.

More arithmetic operators (>=, >, <=, <) are also added so it's easier
to write more expressive programs and tests.

A new flag (defaults to true) is added to the `analyze` function to
optionally verify the result using Spacer. The test framework now
performs unit testing *and* verification testing on analysis results
together.

Test plan:

`dune test program_analysis`

ghstack-source-id: 7553aa7b530d5af9dff4fef9f76f9999dc935e05
Pull Request resolved: https://github.com/JHU-PL-Lab/dde/pull/92