[PA] Stub at Var Non-Local
This commit fixes a major bug in the program analysis where the
execution gets stuck in an infinite loop between the Appl rule and the
Var Non-Local rule. The solution is to stub at Var Non-Local in addition
to Var Local. The program analysis is also now cached based on the
expression being evaluated, the call site stack, and the path condition
that the current execution is under.

Multiplication is added to the language for more varied tests and to
adapt those from benchmarks of other projects.

Test plan:

`dune test program_analysis`

ghstack-source-id: 22f6c17c83b28bf80b9225a5b88be1736d471748
Pull Request resolved: https://github.com/JHU-PL-Lab/dde/pull/97